+++ src/pages/Compare/index.js
@@ -1,0 +1,2 @@
+import Compare from './Compare';
+export default Compare;
